+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - The binary on Windows now statically links to the MSVC Runtime to make it more portable ([#657](https://github.com/JohnnyMorganz/luau-lsp/issues/657))
+- Fixed Roblox types still showing when setting `luau-lsp.platform.type` to something other than `roblox`. ([#668](https://github.com/JohnnyMorganz/luau-lsp/issues/668))
 
 ## [1.31.0] - 2024-07-01